Acreditador/ProyectoV1.0/Julio157Ramirez/ProyectoZip/acreditacion-project_copia/requirements.txt
django==2.1.1
django-import-export==1.0.1